+++ app/src/main/java/com/sopt/now/test/data/mockFriendList.kt
@@ -1,0 +1,56 @@
+package com.sopt.now.test.data
+
+import com.sopt.now.R
+
+var mockFriendList = mutableListOf<Friend>(
+    Friend(
+        profileImage = R.drawable.iv_friend_profile,
+        name = "AAA",
+        selfDescription = "AAA님의 한 줄 소개",
+    ),
+    Friend(
+        profileImage = R.drawable.iv_friend_profile,
+        name = "BBB",
+        selfDescription = "BBB님의 한 줄 소개",
+    ),
+    Friend(
+        profileImage = R.drawable.iv_friend_profile,
+        name = "CCC",
+        selfDescription = "CCC님의 한 줄 소개",
+    ),
+    Friend(
+        profileImage = R.drawable.iv_friend_profile,
+        name = "DDD",
+        selfDescription = "DDD님의 한 줄 소개",
+    ),
+    Friend(
+        profileImage = R.drawable.iv_friend_profile,
+        name = "EEE",
+        selfDescription = "EEE님의 한 줄 소개",
+    ),
+    Friend(
+        profileImage = R.drawable.iv_friend_profile,
+        name = "FFF",
+        selfDescription = "FFF님의 한 줄 소개",
+    ),
+    Friend(
+        profileImage = R.drawable.iv_friend_profile,
+        name = "GGG",
+        selfDescription = "GGG님의 한 줄 소개",
+    ),
+    Friend(
+        profileImage = R.drawable.iv_friend_profile,
+        name = "HHH",
+        selfDescription = "HHH님의 한 줄 소개",
+    ),
+    Friend(
+        profileImage = R.drawable.iv_friend_profile,
+        name = "III",
+        selfDescription = "III님의 한 줄 소개",
+    ),
+    Friend(
+        profileImage = R.drawable.iv_friend_profile,
+        name = "JJJ",
+        selfDescription = "JJJ님의 한 줄 소개",
+    )
+)
